+++ lib/proiel/alignment.rb
@@ -1,0 +1,3 @@
+module PROIEL::Alignment; end
+
+require 'proiel/alignment/builder'